+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/Web.cs
@@ -13,6 +13,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Net;
 using System.Net.Http;
+using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -729,49 +730,186 @@ namespace PnP.Core.Model.SharePoint
 
         private async Task<Guid> GetTenantIdAsync()
         {
-            // in case telemetry is configured, return the tenant id from the globaloptions
-            if (this.PnPContext.GlobalOptions.AADTenantId != Guid.Empty)
-                return this.PnPContext.GlobalOptions.AADTenantId;
+            // in case telemetry is configured, the globaloptions already has a populated tenantid value
+            if (PnPContext.GlobalOptions.AADTenantId == Guid.Empty)
+            {
+                var useOpenIdConfiguration = false;
+#if NET5_0_OR_GREATER
+                useOpenIdConfiguration = RuntimeInformation.RuntimeIdentifier == "browser-wasm";
+#endif
+                await PnPContext.SetAADTenantId(useOpenIdConfiguration).ConfigureAwait(false);
+            }
 
-            await EnsurePropertiesAsync(p => p.Url).ConfigureAwait(false);
-            WebRequest request = WebRequest.Create(this.Url + "/_vti_bin/client.svc");
-            request.Headers.Add("Authorization: Bearer ");
+            if (PnPContext.GlobalOptions.AADTenantId != Guid.Empty)
+            {
+                return PnPContext.GlobalOptions.AADTenantId;
+            }
+            else
+            {
+                return Guid.Empty;                
+            }
+        }
 
+        public ISharePointUser EnsureEveryoneExceptExternalUsers()
+        {
+            return EnsureEveryoneExceptExternalUsersAsync().GetAwaiter().GetResult();
+        }
+
+        public async Task<ISharePointUser> EnsureEveryoneExceptExternalUsersAsync()
+        {
             try
             {
-                await request.GetResponseAsync().ConfigureAwait(false);
+                var tenantId = await GetTenantIdAsync().ConfigureAwait(false);
+                var loginName = $"c:0-.f|rolemanager|spo-grid-all-users/{tenantId}";
+                return await EnsureUserAsync(loginName).ConfigureAwait(false);
             }
-            catch (WebException e)
+            catch(SharePointRestServiceException ex) when (ex.HResult == -2146233088)
             {
-                var bearerResponseHeader = e.Response.Headers["WWW-Authenticate"];
-
-                const string bearer = "Bearer realm=\"";
-                var bearerIndex = bearerResponseHeader.IndexOf(bearer, StringComparison.Ordinal);
-
-                var realmIndex = bearerIndex + bearer.Length;
-
-                if (bearerResponseHeader.Length >= realmIndex + 36)
+                var web = await GetAsync(p=>p.Language).ConfigureAwait(false);
+                string userIdentity = null;
+                switch (web.Language)
                 {
-                    if (Guid.TryParse(bearerResponseHeader.Substring(realmIndex, 36), out Guid realmGuid))
-                    {
-                        return realmGuid;
-                    }
+                    case 1025: // Arabic
+                        userIdentity = "الجميع باستثناء المستخدمين الخارجيين";
+                        break;
+                    case 1069: // Basque
+                        userIdentity = "Guztiak kanpoko erabiltzaileak izan ezik";
+                        break;
+                    case 1026: // Bulgarian
+                        userIdentity = "Всички освен външни потребители";
+                        break;
+                    case 1027: // Catalan
+                        userIdentity = "Tothom excepte els usuaris externs";
+                        break;
+                    case 2052: // Chinese (Simplified)
+                        userIdentity = "除外部用户外的任何人";
+                        break;
+                    case 1028: // Chinese (Traditional)
+                        userIdentity = "外部使用者以外的所有人";
+                        break;
+                    case 1050: // Croatian
+                        userIdentity = "Svi osim vanjskih korisnika";
+                        break;
+                    case 1029: // Czech
+                        userIdentity = "Všichni kromě externích uživatelů";
+                        break;
+                    case 1030: // Danish
+                        userIdentity = "Alle undtagen eksterne brugere";
+                        break;
+                    case 1043: // Dutch
+                        userIdentity = "Iedereen behalve externe gebruikers";
+                        break;
+                    case 1033: // English
+                        userIdentity = "Everyone except external users";
+                        break;
+                    case 1061: // Estonian
+                        userIdentity = "Kõik peale väliskasutajate";
+                        break;
+                    case 1035: // Finnish
+                        userIdentity = "Kaikki paitsi ulkoiset käyttäjät";
+                        break;
+                    case 1036: // French
+                        userIdentity = "Tout le monde sauf les utilisateurs externes";
+                        break;
+                    case 1110: // Galician
+                        userIdentity = "Todo o mundo excepto os usuarios externos";
+                        break;
+                    case 1031: // German
+                        userIdentity = "Jeder, außer externen Benutzern";
+                        break;
+                    case 1032: // Greek
+                        userIdentity = "Όλοι εκτός από εξωτερικούς χρήστες";
+                        break;
+                    case 1037: // Hebrew
+                        userIdentity = "כולם פרט למשתמשים חיצוניים";
+                        break;
+                    case 1081: // Hindi
+                        userIdentity = "बाह्य उपयोगकर्ताओं को छोड़कर सभी";
+                        break;
+                    case 1038: // Hungarian
+                        userIdentity = "Mindenki, kivéve külső felhasználók";
+                        break;
+                    case 1057: // Indonesian
+                        userIdentity = "Semua orang kecuali pengguna eksternal";
+                        break;
+                    case 1040: // Italian
+                        userIdentity = "Tutti tranne gli utenti esterni";
+                        break;
+                    case 1041: // Japanese
+                        userIdentity = "外部ユーザー以外のすべてのユーザー";
+                        break;
+                    case 1087: // Kazakh
+                        userIdentity = "Сыртқы пайдаланушылардан басқасының барлығы";
+                        break;
+                    case 1042: // Korean
+                        userIdentity = "외부 사용자를 제외한 모든 사람";
+                        break;
+                    case 1062: // Latvian
+                        userIdentity = "Visi, izņemot ārējos lietotājus";
+                        break;
+                    case 1063: // Lithuanian
+                        userIdentity = "Visi, išskyrus išorinius vartotojus";
+                        break;
+                    case 1086: // Malay
+                        userIdentity = "Semua orang kecuali pengguna luaran";
+                        break;
+                    case 1044: // Norwegian (Bokmål)
+                        userIdentity = "Alle bortsett fra eksterne brukere";
+                        break;
+                    case 1045: // Polish
+                        userIdentity = "Wszyscy oprócz użytkowników zewnętrznych";
+                        break;
+                    case 1046: // Portuguese (Brazil)
+                        userIdentity = "Todos exceto os usuários externos";
+                        break;
+                    case 2070: // Portuguese (Portugal)
+                        userIdentity = "Todos exceto os utilizadores externos";
+                        break;
+                    case 1048: // Romanian
+                        userIdentity = "Toată lumea, cu excepția utilizatorilor externi";
+                        break;
+                    case 1049: // Russian
+                        userIdentity = "Все, кроме внешних пользователей";
+                        break;
+                    case 10266: // Serbian (Cyrillic, Serbia)
+                        userIdentity = "Сви осим спољних корисника";
+                        break;
+                    case 2074:// Serbian (Latin)
+                        userIdentity = "Svi osim spoljnih korisnika";
+                        break;
+                    case 1051:// Slovak
+                        userIdentity = "Všetci okrem externých používateľov";
+                        break;
+                    case 1060: // Slovenian
+                        userIdentity = "Vsi razen zunanji uporabniki";
+                        break;
+                    case 3082: // Spanish
+                        userIdentity = "Todos excepto los usuarios externos";
+                        break;
+                    case 1053: // Swedish
+                        userIdentity = "Alla utom externa användare";
+                        break;
+                    case 1054: // Thai
+                        userIdentity = "ทุกคนยกเว้นผู้ใช้ภายนอก";
+                        break;
+                    case 1055: // Turkish
+                        userIdentity = "Dış kullanıcılar hariç herkes";
+                        break;
+                    case 1058: // Ukranian
+                        userIdentity = "Усі, крім зовнішніх користувачів";
+                        break;
+                    case 1066: // Vietnamese
+                        userIdentity = "Tất cả mọi người trừ người dùng bên ngoài";
+                        break;
+                }
+
+                if (!string.IsNullOrEmpty(userIdentity))
+                {
+                    return await EnsureUserAsync(userIdentity).ConfigureAwait(false);
                 }
             }
-            return Guid.Empty;
-        }
 
-
-        public ISharePointUser GetEveryoneExceptExternalUsers()
-        {
-            return GetEveryoneExceptExternalUsersAsync().GetAwaiter().GetResult();
-        }
-
-        public async Task<ISharePointUser> GetEveryoneExceptExternalUsersAsync()
-        {
-            var tenantId = await this.GetTenantIdAsync().ConfigureAwait(false);
-            var loginName = $"c:0-.f|rolemanager|spo-grid-all-users/{tenantId}";
-            return await this.EnsureUserAsync(loginName).ConfigureAwait(false);
+            throw new ClientException(ErrorType.Unsupported, PnPCoreResources.Exception_Web_EveyoneExceptUsersCouldNotBeEnsured);
         }
 
         public ISharePointUser EnsureUser(string userPrincipalName)